+++ library/src/main/java/com/getstream/sdk/chat/StreamChat.java
@@ -14,6 +14,9 @@ import com.getstream.sdk.chat.model.Channel;
 import com.getstream.sdk.chat.model.Event;
 import com.getstream.sdk.chat.notifications.NotificationsManager;
 import com.getstream.sdk.chat.notifications.StreamNotificationsManager;
+import com.getstream.sdk.chat.navigation.StreamChatNavigator;
+import com.getstream.sdk.chat.navigation.StreamChatNavigatorImpl;
+import com.getstream.sdk.chat.navigation.ChatNavigationHandler;
 import com.getstream.sdk.chat.rest.User;
 import com.getstream.sdk.chat.rest.core.ApiClientOptions;
 import com.getstream.sdk.chat.rest.core.ChatEventHandler;
@@ -47,6 +50,11 @@ public class StreamChat {
     private static StreamChatStyle chatStyle = new StreamChatStyle.Builder().build();
     private static FontsManager fontsManager;
     private static StreamLogger logger;
+    private static StreamChatNavigator navigator = new StreamChatNavigatorImpl();
+
+    public static StreamChatNavigator getNavigator() {
+        return navigator;
+    }
 
     public static LiveData<OnlineStatus> getOnlineStatus() {
         return onlineStatus;
@@ -107,12 +115,14 @@ public class StreamChat {
     private static void initComponents(String apiKey, ApiClientOptions apiClientOptions) {
         stringsProvider = new StringsProviderImpl(context);
         fontsManager = new FontsManagerImpl(context);
+
         INSTANCE = new Client(apiKey, apiClientOptions, new ConnectionLiveData(StreamChat.context));
 
         onlineStatus = new MutableLiveData<>(OnlineStatus.NOT_INITIALIZED);
         currentUser = new MutableLiveData<>();
         totalUnreadMessages = new MutableLiveData<>();
         unreadChannels = new MutableLiveData<>();
+
 
         INSTANCE.setContext(StreamChat.context);
     }
@@ -254,6 +264,8 @@ public class StreamChat {
         StreamLogger logger = config.getLogger() != null ? config.getLogger() : new StreamChatSilentLogger();
         NotificationsManager notificationsManager = config.getNotificationsManager() != null ? config.getNotificationsManager() : new StreamNotificationsManager();
         return init(config.getApiKey(), options, context, logger, notificationsManager);
+        navigator.setHandler(config.getNavigationHandler());
+        return init(config.getApiKey(), options, context, logger);
     }
 
     public static synchronized boolean init(String apiKey, Context context) {
@@ -298,11 +310,16 @@ public class StreamChat {
         private String apiKey;
         private ApiClientOptions apiClientOptions;
         private StreamLogger logger;
+        private ChatNavigationHandler navigationHandler;
         private NotificationsManager notificationsManager;
 
         public Config(@NonNull Context context, @NonNull String apiKey) {
             StreamChat.context = context;
             this.apiKey = apiKey;
+        }
+
+        public void navigationHandler(ChatNavigationHandler navigationHandler){
+            this.navigationHandler = navigationHandler;
         }
 
         /**
@@ -351,6 +368,10 @@ public class StreamChat {
 
         private StreamLogger getLogger() {
             return logger;
+        }
+
+        private ChatNavigationHandler getNavigationHandler(){
+            return navigationHandler;
         }
 
         private NotificationsManager getNotificationsManager() {