+++ library/src/main/java/com/getstream/sdk/chat/navigation/ChatNavigationHandler.java
@@ -1,0 +1,7 @@
+package com.getstream.sdk.chat.navigation;
+
+import com.getstream.sdk.chat.navigation.destinations.ChatDestination;
+
+public interface ChatNavigationHandler {
+    boolean navigate(ChatDestination destination);
+}