+++ library/src/main/java/com/getstream/sdk/chat/utils/PermissionChecker.java
@@ -3,12 +3,12 @@ package com.getstream.sdk.chat.utils;
 import android.Manifest;
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.net.Uri;
 import android.os.Build;
-import android.provider.Settings;
 import android.widget.Button;
+
+import com.getstream.sdk.chat.StreamChat;
+import com.getstream.sdk.chat.navigation.destinations.AppSettingsDestination;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -87,11 +87,7 @@ public class PermissionChecker {
         alertDialog.setOnShowListener(dialog -> {
             Button button = alertDialog.getButton(AlertDialog.BUTTON_POSITIVE);
             button.setOnClickListener(v -> {
-                Intent intent = new Intent();
-                intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                Uri uri = Uri.fromParts("package", context.getPackageName(), null);
-                intent.setData(uri);
-                context.startActivity(intent);
+                StreamChat.getNavigator().navigate(new AppSettingsDestination(context));
                 alertDialog.dismiss();
             });
         });