+++ sample/src/main/java/io/getstream/chat/example/BaseApplication.java
@@ -1,6 +1,7 @@
 package io.getstream.chat.example;
 
 import android.app.Application;
+import android.widget.Toast;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
@@ -13,6 +14,8 @@ import com.getstream.sdk.chat.logger.StreamChatLogger;
 import com.getstream.sdk.chat.logger.StreamLogger;
 import com.getstream.sdk.chat.logger.StreamLoggerHandler;
 import com.getstream.sdk.chat.logger.StreamLoggerLevel;
+import com.getstream.sdk.chat.navigation.destinations.AttachmentDestination;
+import com.getstream.sdk.chat.navigation.destinations.WebLinkDestination;
 import com.getstream.sdk.chat.model.Event;
 import com.getstream.sdk.chat.notifications.DeviceRegisteredListener;
 import com.getstream.sdk.chat.notifications.NotificationMessageLoadListener;
@@ -29,6 +32,7 @@ import com.google.firebase.messaging.RemoteMessage;
 
 import java.util.Map;
 
+import androidx.annotation.NonNull;
 import io.fabric.sdk.android.Fabric;
 import io.getstream.chat.example.utils.AppDataConfig;
 
@@ -181,6 +185,28 @@ public class BaseApplication extends Application {
         configuration.setApiClientOptions(apiClientOptions);
         configuration.setStyle(style);
         configuration.setLogger(logger);
+        configuration.navigationHandler(destination -> {
+
+            String url = "";
+
+            if (destination instanceof WebLinkDestination) {
+                url = ((WebLinkDestination) destination).url;
+            } else if (destination instanceof AttachmentDestination) {
+                url = ((AttachmentDestination) destination).url;
+            }
+
+            if (url.startsWith("https://your.domain.com")) {
+                Toast.makeText(this, "Custom url handling: " + url, Toast.LENGTH_SHORT).show();
+                // handle/change/update url
+                // open your webview, system browser or your own activity
+                // and return true to override default behaviour
+                return true;
+            } else {
+                return false;
+            }
+
+
+        });
         configuration.setNotificationsManager(notificationsManager);
         StreamChat.init(configuration);